python/requirements.txt
bleach==6.1.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
Flask-HTTPAuth==4.8.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
httplib2==0.22.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
oauth2client==4.1.3
packaging==23.2
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
redis==5.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
urllib3==2.1.0
webencodings==0.5.1
Werkzeug==3.0.1
